+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# RUF r180-Alpha
+# RUF r181-Beta
 ### New
-* Added Pet Happiness Indicator for classic
+* Added unique Pet Happiness Icons (sad face, neutral face, happy face)
 
 ### Known Issues
 * The Options panel settings for Buffs and Debuffs requires a reload to take effect.